refactoring, and build onto existing factory
